requirements.txt
ess-streaming-data-types
kafka-python >= 2.0